requirements.txt
# Core Deep Learning
torch>=2.1.0
transformers>=4.35.0
accelerate>=0.24.0
bitsandbytes>=0.41.0  # For quantization
peft>=0.6.0  # For LoRA

# Reinforcement Learning
gymnasium>=0.29.0
stable-baselines3>=2.1.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
hypothesis>=6.88.0  # For property-based testing

# Logging and Monitoring
tensorboard>=2.14.0
wandb>=0.15.0
tqdm>=4.66.0

# Configuration
pyyaml>=6.0
hydra-core>=1.3.0

# Utilities
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0

# Code Quality
black>=23.0.0
flake8>=6.1.0
mypy>=1.5.0
